% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instar_report.R
\name{autoplot.instar_report}
\alias{autoplot.instar_report}
\title{Return a report's figure as a plot object}
\usage{
\method{autoplot}{instar_report}(object, value_wrap = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{instar_report}.}

\item{value_wrap}{Approximate characters per line for the item text.
Defaults to the value stored on the report by \code{\link[=instar_report]{instar_report()}}.}

\item{...}{Unused.}
}
\value{
A patchwork object.
}
\description{
The \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} method for \code{instar_report} objects. Unlike
\link[=plot.instar_report]{plot()}, which draws to the device, this returns
the patchwork composition so you can modify it before rendering:
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{autoplot(rep) + patchwork::plot_annotation(caption = "Figure S1")
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
rep <- instar_report(instar_template(), list(title = "T", authors = "A"))
p <- autoplot(rep)
}

}
