% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instar.R
\name{.corpus_names}
\alias{.corpus_names}
\title{Name a corpus, preferring DOIs over file names}
\usage{
.corpus_names(reports, labels)
}
\description{
Name a corpus, preferring DOIs over file names
}
\keyword{internal}
