% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instar.R
\name{.read_one}
\alias{.read_one}
\title{Read one sheet of any supported format into a report}
\usage{
.read_one(path, strict = FALSE)
}
\description{
Read one sheet of any supported format into a report
}
\keyword{internal}
