% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instar.R
\name{.warn_version_mix}
\alias{.warn_version_mix}
\title{Warn when a corpus mixes framework versions}
\usage{
.warn_version_mix(x)
}
\description{
Warn when a corpus mixes framework versions
}
\keyword{internal}
