% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{new_instar_audit}
\alias{new_instar_audit}
\title{Construct an instar_audit from its parts}
\usage{
new_instar_audit(long, studies, meta, corpus)
}
\description{
Construct an instar_audit from its parts
}
\keyword{internal}
