% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{new_instar_corpus}
\alias{new_instar_corpus}
\title{Build a corpus from a list of reports}
\usage{
new_instar_corpus(x)
}
\arguments{
\item{x}{A list of \code{instar_report} objects.}
}
\value{
An \code{instar_corpus}.
}
\description{
Build a corpus from a list of reports
}
\keyword{internal}
