% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instar_report.R
\name{plot.instar_report}
\alias{plot.instar_report}
\title{Draw the standardised INSTAR figure}
\usage{
\method{plot}{instar_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{instar_report}.}

\item{...}{Passed to \link[=autoplot.instar_report]{autoplot()}, notably
\code{value_wrap}.}
}
\value{
\code{x}, invisibly.
}
\description{
Renders a report to the current graphics device. This is the usual way
to look at a report. To capture the figure as an object you can modify
or compose with, use \link[=autoplot.instar_report]{autoplot()} instead; to
write it straight to a file, use \code{\link[=save_figure]{save_figure()}}.
}
\examples{
\dontrun{
rep <- instar_report(instar_template(), list(title = "T", authors = "A"))
plot(rep)
}

}
