% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{print.instar_audit}
\alias{print.instar_audit}
\title{Print an audit}
\usage{
\method{print}{instar_audit}(x, n = 5, ...)
}
\arguments{
\item{x}{An \code{instar_audit}.}

\item{n}{Number of worst-reported items to list. Defaults to 5.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an audit
}
