% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{print.instar_corpus}
\alias{print.instar_corpus}
\title{Print an INSTAR corpus}
\usage{
\method{print}{instar_corpus}(x, ...)
}
\arguments{
\item{x}{An \code{instar_corpus}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an INSTAR corpus
}
